completed exercise 4.21

+++ osa4/bloglist/controllers/blogs.js
@@ -51,8 +51,18 @@ blogsRouter.post('/', async (request, response, next) => {
 blogsRouter.delete('/:id', async (request, response, next) => {
   const blogId = request.params.id
   try {
-    await Blog.findByIdAndRemove(blogId)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(blogId)
+    if (blog.user.toString() === decodedToken.id) {
+      await Blog.findByIdAndRemove(blogId)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'user cannot delete this blog' })
+    }
   } catch (exception) {
     next(exception)
   }
